+++ src/Producer/Adapter/ArrayAdapter.php
@@ -3,9 +3,9 @@
 namespace FiiSoft\Jackdaw\Producer\Adapter;
 
 use FiiSoft\Jackdaw\Internal\Item;
-use FiiSoft\Jackdaw\Producer\Producer;
+use FiiSoft\Jackdaw\Producer\BaseProducer;
 
-final class ArrayAdapter implements Producer
+final class ArrayAdapter extends BaseProducer
 {
     private array $source;