+++ src/Producer/Resource/PDOStatementAdapter.php
@@ -3,9 +3,9 @@
 namespace FiiSoft\Jackdaw\Producer\Resource;
 
 use FiiSoft\Jackdaw\Internal\Item;
-use FiiSoft\Jackdaw\Producer\Producer;
+use FiiSoft\Jackdaw\Producer\BaseProducer;
 
-final class PDOStatementAdapter implements Producer
+final class PDOStatementAdapter extends BaseProducer
 {
     private \PDOStatement $statement;
     private ?int $fetchMode;